Move remaining LSP logic from Project to LspStore
